[added] color implementation

+++ app/src/main/java/app/com/huelightsalarm/interfaces/LightsModifier.java
@@ -1,6 +1,6 @@
 package app.com.huelightsalarm.interfaces;
 
 public interface LightsModifier {
-    void setLightColor(String id, int hue, int saturation, int brightness);
+    void setLightColor(String id, float hue, float saturation, float brightness);
     void setLightState(String id, boolean isOn);
 }
